+++ CHANGELOG.md
@@ -8,6 +8,7 @@ IMPROVEMENTS:
    to find all failing service instances on all DCs very easily.
  * Do not report timestamps in metrics.erb as it is toxic for Prometheus
  * in metrics.erb, add all net_info statistics to observe consul-templaterb itself
+ * new flag -l <error|info|debug> to control verbosity of messages
 
 ## 1.9.9 (January 18, 2019)